Extract TotalBalances component
